Metric url is now not requred

+++ app/models/schemas.py
@@ -8,8 +8,8 @@ class StakeholderResponse(BaseModel):
     did: str
     name: str
     created_at: datetime
-    probabilistic_trust: float
-    deterministic_trust: float
+    probabilistic_trust: int
+    deterministic_trust: int
 
 class AllStakeholdersResponse(BaseModel):
     stakeholders: List[StakeholderResponse]
